+++ modules/Attendance/DuplicateAttendance.php
@@ -15,11 +15,6 @@ if ( ! empty( $_REQUEST['mp_arr'] ) )
 	$last_mp = $mp;
 }
 
-//Widgets('course');
-//Widgets('gpa');
-//Widgets('class_rank');
-//Widgets('letter_grade');
-
 if ( $_REQUEST['delete'] === 'true' )
 {
 	//DeletePrompt(_('Duplicate Attendance Record'));
@@ -146,22 +141,24 @@ if ( isset( $_REQUEST['search_modfunc'] )
 		//$extra['WHERE'] .= " AND ssm.student_id=s.student_id AND ap.STUDENT_ID=s.STUDENT_ID AND ap.COURSE_PERIOD_ID = cp.COURSE_PERIOD_ID AND ('".DBDate()."' BETWEEN ssm.START_DATE AND ssm.END_DATE OR ssm.END_DATE IS NULL) ";
 		//$extra['WHERE'] .= " AND ssm.student_id=s.student_id AND ap.STUDENT_ID=s.STUDENT_ID AND ap.COURSE_PERIOD_ID = cp.COURSE_PERIOD_ID ";
 
+		// @since 10.2.1 SQL handle case when student dropped and then later re-enrolled in course.
 		$extra['WHERE'] = " AND ap.STUDENT_ID=s.STUDENT_ID
 			AND sc.STUDENT_ID=s.STUDENT_ID
 			AND ap.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID
 			AND ap.COURSE_PERIOD_ID=sc.COURSE_PERIOD_ID
-			AND sc.END_DATE>'1999-01-01' ";
+			AND sc.END_DATE>'1999-01-01'
+			AND NOT EXISTS(SELECT 1 FROM schedule
+			WHERE COURSE_PERIOD_ID=sc.COURSE_PERIOD_ID
+			AND STUDENT_ID=sc.STUDENT_ID
+			AND END_DATE IS NULL) ";
 
 		$extra['ORDER_BY'] = ' STUDENT_ID,COURSE_PERIOD_ID,SCHOOL_DATE';
-
-		Widgets( 'course' );
-		Widgets( 'gpa' );
-		Widgets( 'class_rank' );
-		Widgets( 'letter_grade' );
 
 		$pageresult1 = GetStuList( $extra );
 
 		$totalrows = 0;
+
+		$studentid2 = 0;
 
 		foreach ( (array) $pageresult1 as $rr )
 		{
@@ -181,7 +178,7 @@ if ( isset( $_REQUEST['search_modfunc'] )
 				$afterr = "Y";
 			}
 
-			if ( isset( $studentid2 )
+			if ( $studentid2
 				&& $studentidr == $studentid2
 				&& $courseidr == $courseid2
 				&& $schooldater == $schooldate2
@@ -220,18 +217,18 @@ if ( isset( $_REQUEST['search_modfunc'] )
 		//$extra['WHERE'] .= " AND ssm.student_id=s.student_id AND ap.STUDENT_ID=s.STUDENT_ID AND ap.COURSE_PERIOD_ID = cp.COURSE_PERIOD_ID AND ('".DBDate()."' BETWEEN ssm.START_DATE AND ssm.END_DATE OR ssm.END_DATE IS NULL) ";
 		//$extra['WHERE'] .= " AND ssm.student_id=s.student_id AND ap.STUDENT_ID=s.STUDENT_ID AND ap.COURSE_PERIOD_ID = cp.COURSE_PERIOD_ID ";
 
+		// @since 10.2.1 SQL handle case when student dropped and then later re-enrolled in course.
 		$extra['WHERE'] = " AND ap.STUDENT_ID=s.STUDENT_ID
 			AND sc.STUDENT_ID=s.STUDENT_ID
 			AND ap.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID
 			AND ap.COURSE_PERIOD_ID=sc.COURSE_PERIOD_ID
-			AND sc.END_DATE>'1999-01-01' ";
+			AND sc.END_DATE>'1999-01-01'
+			AND NOT EXISTS(SELECT 1 FROM schedule
+			WHERE COURSE_PERIOD_ID=sc.COURSE_PERIOD_ID
+			AND STUDENT_ID=sc.STUDENT_ID
+			AND END_DATE IS NULL) ";
 
 		$extra['ORDER_BY'] = ' STUDENT_ID,COURSE_PERIOD_ID,SCHOOL_DATE';
-
-		Widgets( 'course' );
-		Widgets( 'gpa' );
-		Widgets( 'class_rank' );
-		Widgets( 'letter_grade' );
 
 		$result1 = GetStuList( $extra );